third_party/WebKit/LayoutTests/fast/dom/createElementNS-namespace-errors-expected.txt

Test that calling createElementNS() throws an error when http://www.w3.org/TR/dom/#dom-document-createelementns says it should.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElementNS(null, 'foo:bar', 'baz') threw exception NamespaceError: Failed to execute 'createElementNS' on 'Document': The namespace URI provided ('') is not valid for the qualified name provided ('foo:bar')..
PASS document.createElementNS('http://www.w3.org/XML/1998/namespace', 'xml:abc', 'foo') did not throw exception.
PASS document.createElementNS('http://www.w3.org/not-XML/1998/namespace', 'xml:abc', 'foo') threw exception NamespaceError: Failed to execute 'createElementNS' on 'Document': The namespace URI provided ('http://www.w3.org/not-XML/1998/namespace') is not valid for the qualified name provided ('xml:abc')..
PASS document.createElementNS('http://www.w3.org/2000/xmlns/', 'xmlns', 'http://wwww.example.org') did not throw exception.
PASS document.createElementNS('http://www.w3.org/2000/not-xmlns/', 'xmlns', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'createElementNS' on 'Document': The namespace URI provided ('http://www.w3.org/2000/not-xmlns/') is not valid for the qualified name provided ('xmlns')..
PASS document.createElementNS('http://www.w3.org/2000/xmlns/', 'xmlns:abc', 'http://wwww.example.org') did not throw exception.
PASS document.createElementNS('http://www.w3.org/2000/not-xmlns/', 'xmlns:abc', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'createElementNS' on 'Document': The namespace URI provided ('http://www.w3.org/2000/not-xmlns/') is not valid for the qualified name provided ('xmlns:abc')..
PASS document.createElementNS('http://www.w3.org/2000/xmlns/', 'badprefix:xmlns', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'createElementNS' on 'Document': The namespace URI provided ('http://www.w3.org/2000/xmlns/') is not valid for the qualified name provided ('badprefix:xmlns')..
PASS document.createElementNS('http://www.w3.org/2000/xmlns/', 'notxmlns', 'http://wwww.example.org') threw exception NamespaceError: Failed to execute 'createElementNS' on 'Document': The namespace URI provided ('http://www.w3.org/2000/xmlns/') is not valid for the qualified name provided ('notxmlns')..
PASS successfullyParsed is true

TEST COMPLETE
